requirements.txt
alembic==1.4.1
aniso8601==9.0.1
appdirs==1.4.4
attrs==20.3.0
azure-core==1.12.0
azure-storage-blob==12.8.0
boto3==1.17.33
botocore==1.20.33
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
coverage==5.5
cryptography==3.4.6
databricks-cli==0.14.3
decorator==4.4.2
dill==0.3.3
distlib==0.3.1
docker==4.4.4
dotmap==1.3.23
elastic-apm==6.0.0
elasticsearch==7.10.0
entrypoints==0.3
faiss-cpu==1.6.3
farm==0.6.2
farm-haystack==0.7.0
fastapi==0.63.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
gitdb==4.0.5
GitPython==3.1.14
greenlet==1.0.0
gunicorn==20.0.4
h11==0.12.0
httptools==0.1.1
idna==2.10
importlib-metadata==3.7.3
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
langdetect==1.0.8
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
mlflow==1.13.1
more-itertools==8.7.0
msrest==0.6.21
networkx==2.5
nltk==3.5
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pdftotext==2.1.5
pluggy==0.13.1
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
protobuf==3.15.6
psutil==5.8.0
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pydantic==1.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.10
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
querystring-parser==1.2.4
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.6
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
sentencepiece==0.1.95
seqeval==1.2.2
six==1.15.0
sklearn==0.0
smmap==3.0.5
SQLAlchemy==1.4.2
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
starlette==0.13.6
tabulate==0.8.9
threadpoolctl==2.1.0
tika==1.24
tokenizers==0.9.4
toml==0.10.2
torch==1.7.1
tox==3.23.0
tqdm==4.59.0
transformers==4.1.1
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
virtualenv==20.4.3
websocket-client==0.58.0
Werkzeug==0.16.1
zipp==3.4.1
